+++ src/generation/claude_assistant.py
@@ -7,13 +7,10 @@ import tiktoken
 from anthropic.types import Message
 from anthropic.types.beta.prompt_caching import PromptCachingBetaMessage
 
-from src.generation.query_generator import QueryGenerator
 from src.generation.tool_definitions import tool_manager
 from src.utils.config import ANTHROPIC_API_KEY
-from src.utils.decorators import error_handler
-from src.utils.logger import setup_logger
-
-logger = setup_logger("claude_assistant", "claude_assistant.log")
+from src.utils.decorators import anthropic_error_handler, base_error_handler
+from src.utils.logger import logger
 
 
 class ConversationMessage:
@@ -36,7 +33,7 @@ class ConversationHistory:
         self.messages: list[ConversationMessage] = []
         self.total_tokens = 0
         self.tokenizer = tiktoken.get_encoding(tokenizer)
-        self.logger = setup_logger(__name__, "claude_assistant.log")
+        self.logger = logger.get_logger(__name__)
 
     def add_message(self, role: str, content: str | list[dict[str, Any]]) -> None:
         message = ConversationMessage(role, content)
@@ -98,7 +95,7 @@ class ClaudeAssistant:
         self.client = None
         self.api_key = api_key
         self.model_name = model_name
-        self.logger = logger
+        self.logger = logger.get_logger(__name__)
         self.base_system_prompt = """
         You are an advanced AI assistant with access to various tools, including a powerful RAG (Retrieval Augmented
         Generation) system. Your primary function is to provide accurate, relevant, and helpful information to users
@@ -158,18 +155,17 @@ class ClaudeAssistant:
         self.tools: list[dict[str, Any]] = []
         self.extra_headers = {"anthropic-beta": "prompt-caching-2024-07-31"}
         self.retriever = None
-        self.query_generator = QueryGenerator()
 
         self._init()
 
-    @error_handler(logger)
+    @anthropic_error_handler(logger)
     def _init(self):
         self.client = anthropic.Anthropic(api_key=self.api_key, max_retries=2)  # default number of re-tries
         self.conversation_history = ConversationHistory()
 
         self.tools = self.tool_manager.get_all_tools()  # Get all tools as a list of dicts
 
-    @error_handler(logger)
+    @base_error_handler(logger)
     def update_system_prompt(self, document_summaries: list[dict[str, Any]]):
         """
         :param document_summaries: List of dictionaries containing summary information to be incorporated into
@@ -181,11 +177,11 @@ class ClaudeAssistant:
         self.system_prompt = self.base_system_prompt.format(document_summaries=summaries_text)
         self.logger.debug(f"Updated system prompt: {self.system_prompt}")
 
-    @error_handler(logger)
+    @base_error_handler(logger)
     def cached_system_prompt(self) -> list[dict[str, Any]]:
         return [{"type": "text", "text": self.system_prompt, "cache_control": {"type": "ephemeral"}}]
 
-    @error_handler(logger)
+    @base_error_handler(logger)
     def cached_tools(self) -> list[dict[str, Any]]:
         return [
             {
@@ -197,7 +193,7 @@ class ClaudeAssistant:
             for tool in self.tools
         ]
 
-    @error_handler(logger)
+    @base_error_handler(logger)
     def preprocess_user_input(self, input_text: str) -> str:
         # Remove any leading/trailing whitespace
         cleaned_input = input_text.strip()
@@ -205,7 +201,7 @@ class ClaudeAssistant:
         cleaned_input = " ".join(cleaned_input.split())
         return cleaned_input
 
-    @error_handler(logger)
+    @anthropic_error_handler(logger)
     def get_response(self, user_input: str, stream: bool = True) -> Generator[str, None, None] | str:
         """
         :param stream: controls whether output is streamed or not
@@ -222,19 +218,16 @@ class ClaudeAssistant:
             assistant_response = self.not_stream_response(user_input)
             return assistant_response
 
-    @error_handler(logger)
+    @anthropic_error_handler(logger)
     def stream_response(self, user_input: str) -> Generator[str, None, None]:
-        MAX_ITERATIONS = 2
-        iteration = 0
+        # iteration = 0
         user_input = self.preprocess_user_input(user_input)
         self.conversation_history.add_message(role="user", content=user_input)
         self.logger.debug(
             f"Printing conversation history for debugging: {self.conversation_history.get_conversation_history()}"
         )
 
-        while iteration < MAX_ITERATIONS:
-            iteration += 1
-            self.logger.debug(f"Stream response iteration {iteration}")
+        while True:
 
             try:
                 messages = self.conversation_history.get_conversation_history()
@@ -256,17 +249,15 @@ class ClaudeAssistant:
                             if event.content_block.type == "tool_use":
                                 self.logger.debug(f"Tool use detected: {event.content_block.name}")
                                 yield {"type": "tool_use", "tool": event.content_block.name}
-                                # yield f"Using {event.content_block.name} tool"
-
                         elif event.type == "message_stop":
                             self.logger.debug("===== Stream message ended =====")
 
                 # Get the final message after consuming the entire stream
-                final_message = stream.get_final_message()
-                self._process_assistant_response(final_message)
+                assistant_response = stream.get_final_message()
+                self._process_assistant_response(assistant_response)
 
                 # Handle tool use if present in the final message
-                tool_use_block = next((block for block in final_message.content if block.type == "tool_use"), None)
+                tool_use_block = next((block for block in assistant_response.content if block.type == "tool_use"), None)
                 if tool_use_block:
                     tool_result = self.handle_tool_use(tool_use_block.name, tool_use_block.input, tool_use_block.id)
                     self.logger.debug(f"Tool result: {tool_result}")
@@ -280,7 +271,7 @@ class ClaudeAssistant:
                 self.conversation_history.remove_last_message()
                 raise Exception(f"An error occurred: {str(e)}") from e
 
-    @error_handler(logger)
+    @anthropic_error_handler(logger)
     def not_stream_response(self, user_input: str) -> str:
         user_input = self.preprocess_user_input(user_input)
         self.conversation_history.add_message(role="user", content=user_input)
@@ -289,7 +280,7 @@ class ClaudeAssistant:
         )
 
         try:
-            while True:  # TODO: implement max_iter to ensure we exit the loop if max tool use is used
+            while True:
                 messages = self.conversation_history.get_conversation_history()
                 response = self.client.beta.prompt_caching.messages.create(
                     messages=messages,
@@ -302,7 +293,7 @@ class ClaudeAssistant:
                 # tool use
                 if response.stop_reason == "tool_use":
                     assistant_response = self._process_assistant_response(response)  # save the first response
-                    print(f"Assistant: Using a 🔨tool: {assistant_response}")  # TODO: Re-factor to assistnat message
+                    print(f"Assistant: Using a 🔨tool: {assistant_response}")
 
                     tool_use = next(block for block in response.content if block.type == "tool_use")
 
@@ -319,7 +310,7 @@ class ClaudeAssistant:
             self.conversation_history.remove_last_message()
             raise Exception(f"An error occurred: {str(e)}") from e
 
-    @error_handler(logger)
+    @base_error_handler(logger)
     def _process_assistant_response(self, response: PromptCachingBetaMessage | Message) -> str:
         """
         Process the assistant's response by saving the message and updating the token count.
@@ -338,7 +329,7 @@ class ClaudeAssistant:
 
         return response.content[0].text
 
-    @error_handler(logger)
+    @base_error_handler(logger)
     def handle_tool_use(self, tool_name: str, tool_input: dict[str, Any], tool_use_id: str) -> dict[str, Any] | str:
 
         try:
@@ -351,7 +342,7 @@ class ClaudeAssistant:
                         {
                             "type": "tool_result",
                             "tool_use_id": tool_use_id,
-                            "content": f"Here is context retrieved by a RAG system: \n\n{search_results}\n\n."
+                            "content": f"Here is context retrieved by RAG search: \n\n{search_results}\n\n."
                             f"Now please try to answer my original request.",
                         }
                     ],
@@ -359,12 +350,16 @@ class ClaudeAssistant:
 
                 # save message to conversation history
                 self.conversation_history.add_message(**tool_result)
+                self.logger.debug(
+                    f"Debugging conversation history after tool use: "
+                    f"{self.conversation_history.get_conversation_history()}"
+                )
                 return tool_result
         except Exception as e:
             self.logger.error(f"Error executing tool {tool_name}: {str(e)}")
             return f"Error: {str(e)}"
 
-    @error_handler(logger)
+    @anthropic_error_handler(logger)
     def formulate_rag_query(self, recent_conversation_history: list[dict[str, Any]], important_context: str) -> str:
         """ "
         Generates a rag search query based on recent conversation history and important context generated by AI
@@ -422,13 +417,13 @@ class ClaudeAssistant:
         rag_query = response.content[0].text.strip()
         return rag_query
 
-    @error_handler(logger)
+    @base_error_handler(logger)
     def get_recent_context(self, n_messages: int = 6) -> list[dict[str, str]]:
         """Retrieves last 6 messages (3 user messages + 3 assistant messages)"""
         recent_messages = self.conversation_history.messages[-n_messages:]
         return [msg.to_dict() for msg in recent_messages]
 
-    @error_handler(logger)
+    @base_error_handler(logger)
     def use_rag_search(self, tool_input: dict[str, Any]) -> list[str]:
         # Get recent conversation context (last n messages for each role)
         recent_conversation_history = self.get_recent_context()
@@ -439,8 +434,8 @@ class ClaudeAssistant:
         # prepare queries for search
         rag_query = self.formulate_rag_query(recent_conversation_history, important_context)
         self.logger.debug(f"Using this query for RAG search: {rag_query}")
-        multiple_queries = self.query_generator.generate_multi_query(rag_query)
-        combined_queries = self.query_generator.combine_queries(rag_query, multiple_queries)
+        multiple_queries = self.generate_multi_query(rag_query)
+        combined_queries = self.combine_queries(rag_query, multiple_queries)
 
         # get ranked search results
         results = self.retriever.retrieve(rag_query, combined_queries)
@@ -452,7 +447,7 @@ class ClaudeAssistant:
 
         return preprocessed_results
 
-    @error_handler(logger)
+    @anthropic_error_handler(logger)
     def generate_document_summary(self, chunks: list[dict[str, Any]]) -> dict[str, Any]:
         # Aggregate metadata
         unique_urls = {chunk["metadata"]["source_url"] for chunk in chunks}
@@ -519,7 +514,7 @@ class ClaudeAssistant:
     def _format_content_samples(self, samples: list[str]) -> str:
         return "\n\n".join(f"Sample {i + 1}:\n{sample}" for i, sample in enumerate(samples))
 
-    @error_handler(logger)
+    @base_error_handler(logger)
     def preprocess_ranked_documents(self, ranked_documents: dict[str, Any]) -> list[str]:
         """
         Converts ranked documents into a structured string for passing to the Claude API.
@@ -537,3 +532,40 @@ class ClaudeAssistant:
             preprocessed_context.append(formatted_document)
 
         return preprocessed_context
+
+    @anthropic_error_handler(logger)
+    def generate_multi_query(self, query: str, model: str = None, n_queries: int = 5) -> list[str]:
+        prompt = f"""
+            You are an AI assistant whose task is to generate multiple queries as part of a RAG system.
+            You are helping users retrieve relevant information from a vector database.
+            For the given user question, formulate up to {n_queries} related, relevant questions to assist in
+            finding the information.
+
+            Requirements to follow:
+            - Do NOT include any other text in your response except for 3 queries, each on a separate line.
+            - Provide concise, single-topic questions (without compounding sentences) that cover various aspects of
+            the topic.
+            - Ensure each question is complete and directly related to the original inquiry.
+            - List each question on a separate line without numbering.
+            """
+        if model is None:
+            model = self.model_name
+
+        message = self.client.messages.create(
+            model="claude-3-5-sonnet-20240620",
+            max_tokens=1024,
+            system=prompt,
+            messages=[{"role": "user", "content": query}],
+        )
+
+        content = message.content[0].text
+        content = content.split("\n")
+        return content
+
+    @base_error_handler(logger)
+    def combine_queries(self, user_query: str, generated_queries: list[str]) -> list[str]:
+        """
+        Combines user query and generated queries into a list, removing any empty queries.
+        """
+        combined_queries = [query for query in [user_query] + generated_queries if query.strip()]
+        return combined_queries