+++ src/vector_storage/vector_db.py
@@ -8,19 +8,9 @@ import chromadb.utils.embedding_functions as embedding_functions
 import cohere
 
 from src.generation.claude_assistant import ClaudeAssistant
-from src.utils.config import (
-    CHROMA_DB_DIR,
-    COHERE_API_KEY,
-    LOG_DIR,
-    OPENAI_API_KEY,
-    PROCESSED_DATA_DIR,
-    VECTOR_STORAGE_DIR,
-)
-from src.utils.decorators import error_handler
-from src.utils.logger import setup_logger
-
-# Set up logger
-logger = setup_logger(__name__, os.path.join(LOG_DIR, "app.log"))
+from src.utils.config import CHROMA_DB_DIR, COHERE_API_KEY, OPENAI_API_KEY, PROCESSED_DATA_DIR, VECTOR_STORAGE_DIR
+from src.utils.decorators import base_error_handler
+from src.utils.logger import logger
 
 
 class DocumentProcessor:
@@ -67,7 +57,7 @@ class VectorDB:
             f"{self.collection.count()} documents (chunks)"
         )
 
-    @error_handler(logger)
+    @base_error_handler(logger)
     def _load_existing_summaries(self):
         summaries_file = os.path.join(VECTOR_STORAGE_DIR, "document_summaries.json")
         if os.path.exists(summaries_file):
@@ -109,7 +99,7 @@ class VectorDB:
 
         return {"ids": ids, "documents": documents, "metadatas": metadatas}
 
-    @error_handler(logger)
+    @base_error_handler(logger)
     def add_documents(self, json_data: list[dict], claude_assistant: ClaudeAssistant, file_name: str) -> str | None:
         processed_docs = self.prepare_documents(json_data)
 
@@ -122,7 +112,7 @@ class VectorDB:
 
         if all_exist:
             logger.info(
-                f"All documents from {file_name}already loaded. Proceeding to generate or load summary for "
+                f"All documents from {file_name} already loaded. Proceeding to generate or load summary for "
                 f"the "
                 f"entire file."
             )
@@ -155,7 +145,7 @@ class VectorDB:
         self._save_summaries()
         return summary
 
-    @error_handler(logger)
+    @base_error_handler(logger)
     def _save_summaries(self):
         summaries_file = os.path.join(VECTOR_STORAGE_DIR, "document_summaries.json")
         try:
@@ -164,11 +154,11 @@ class VectorDB:
         except Exception as e:
             logger.error(f"Failed to save document summaries: {e}")
 
-    @error_handler(logger)
+    @base_error_handler(logger)
     def get_document_summaries(self) -> list[str]:
         return list(self.document_summaries.values())
 
-    @error_handler(logger)
+    @base_error_handler(logger)
     def check_documents_exist(self, document_ids: list[str]) -> tuple[bool, list[str]]:
         """Checks if chunks are already added to the database based on chunk ids"""
         try:
@@ -188,7 +178,7 @@ class VectorDB:
             logger.error(f"Error checking document existence: {e}")
             return False, document_ids
 
-    @error_handler(logger)
+    @base_error_handler(logger)
     def query(self, user_query: str | list[str], n_results: int = 5):
         """
         Handles both a single query and multiple queris
@@ -202,7 +192,7 @@ class VectorDB:
         )
         return search_results
 
-    @error_handler(logger)
+    @base_error_handler(logger)
     def reset_database(self):
         self.client.delete_collection(self.collection_name)
         self.collection = self.client.create_collection(